update prpitise from float to int

+++ Models/Location.cs
@@ -2,7 +2,7 @@
 {
     public class Location
     {
-        public float X { get; set; }
-        public float Y { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
     }
 }
